% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{date_ts2date}
\alias{date_ts2date}
\title{Conversion d'une date du format TS au format date}
\usage{
date_ts2date(date_ts, frequency_ts)
}
\arguments{
\item{date_ts}{un vecteur numérique, de préférence \code{integer}, au format
\code{AAAA}, \code{c(AAAA, MM)} ou \code{c(AAAA, TT)}}

\item{frequency_ts}{un entier qui vaut \code{4L} (ou \code{4.0}) pour les séries
trimestrielles et \code{12L} (ou \code{12.0}) pour les séries mensuelles.}
}
\value{
En sortie, la fonction retourne un objet de type Date (atomic) de
longueur 1 qui correspond à l'objet \code{date_ts}.
}
\description{
Conversion d'une date du format TS au format date
}
\examples{

date_ts2date(date_ts = c(2020L, 11L), frequency_ts = 12L)
date_ts2date(date_ts = c(1995L, 2L), frequency_ts = 4L)

}
