% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ev_pib}
\alias{ev_pib}
\title{Évolution du PIB français jusqu'au T1 2022}
\format{
Un ts unidimensionnel :
\describe{
\item{start}{le ts commence au T1 1970 mais la série de PIB ne commence qu'au
T2 1980.}
\item{end}{le ts finit au T3 2022 mais la série de PIB finit au T1 2022.}
\item{frequency_ts}{la fréquence est trimestrielle}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2830547}
}
\usage{
ev_pib
}
\description{
Ce jeu de données contient une série \code{ts} de l'évolution trimestrielle du
produit intérieur brut français.
Toutes les infos complémentaires sur cette série se trouve sur la page de la
\href{https://www.insee.fr/fr/statistiques/2830547}{publication} sur le site
de l'\href{https://www.insee.fr/fr/accueil}{Insee}.
}
\keyword{datasets}
