% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_manipulation.R
\name{first_date}
\alias{first_date}
\title{Première date non NA}
\usage{
first_date(series)
}
\arguments{
\item{series}{un objet ts unidimensionnel conforme aux règles de assert_ts}
}
\value{
En sortie, la fonction retourne un objet au format \code{date_ts} (\code{AAAA},
\code{c(AAAA, MM)} ou \code{c(AAAA, TT)})
}
\description{
Cette fonction calcule la première date pour laquelle l'objet
\code{series} ne vaut pas NA.
}
\details{
La date retournée en output est au format \code{date_ts}. Si l'objet
\code{series} ne contient que des NAs, la fonction retourne une erreur.
}
\examples{

ts1 <- ts(c(NA, NA, NA, 1:10, NA), start = 2000, frequency = 12)
ts2 <- ts(c(1:10, NA), start = 2020, frequency = 4)

stats::start(ts1)
first_date(ts1)

stats::start(ts1)
first_date(ts2)

}
\seealso{
\code{last_date}
}
