% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{set_value_ts}
\alias{set_value_ts}
\title{Change certaines valeurs d'un ts}
\usage{
set_value_ts(series, date_ts, replacement)
}
\arguments{
\item{series}{un objet ts unidimensionnel conforme aux règles de assert_ts}

\item{date_ts}{un vecteur numérique, de préférence \code{integer} au format
\code{AAAA}, \code{c(AAAA, MM)} ou \code{c(AAAA, TT)}}

\item{replacement}{un vecteur de même type que le ts \code{series}}
}
\value{
En sortie, la fonction retourne une copie de l'objet \code{series} modifié
avec les valeurs de \code{replacement} imputés à partir de la date \code{date_ts}.
}
\description{
La fonction \code{set_value_ts} modifie la ou les valeurs d'un objet
ts à une date donnée.
}
\examples{
set_value_ts(
    series = ev_pib,
    date_ts = c(2021L, 2L),
    replacement = c(1, 2, 3)
)

}
