% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{substr_year}
\alias{substr_year}
\title{Retire une année à une date}
\usage{
substr_year(date, n = 1L)
}
\arguments{
\item{date}{un objet de type Date}

\item{n}{un entier}
}
\value{
En sortie, la fonction retourne un objet de type Date (atomic) de
longueur 1.
}
\description{
La fonction \code{substr_year} retire \code{n} annnée(s) à une date.
}
\examples{

substr_year(as.Date("2000-02-29"), n = 1L)
substr_year(as.Date("2000-02-29"), n = 3L)
substr_year(as.Date("2000-02-29"), n = 4L)
substr_year(as.Date("2000-02-29"), n = 16L)

substr_year(as.Date("2023-01-25"), n = 10L)
substr_year(as.Date("2022-11-01"), n = 3L)

}
